FlaskServer/requirements.txt
fastai==1.0.51
tensorboardX==1.6
ffmpeg
boto3
ffmpeg-python==0.1.17
youtube-dl>=2019.4.17
opencv-python>=3.3.0.10
pillow
flask==1.1.3
celery
# Fix for: https://github.com/celery/kombu/issues/1063
kombu==4.5.0
flower
flask-restx
redis
gunicorn
torch==1.5.0
torchvision==0.6.0